# octbase-api/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Octbase API
  version: "0.4.0"
  description: |
    REST API for Octbase — a focused project-management tool (tasks, boards,
    releases, sprints, wiki pages, SCM integration) for a single, invitation-only
    team.

    **Authentication:** All routes require a JWT **Bearer** access token in the
    `Authorization` header, except the public auth/invitation routes, the
    HMAC-authenticated webhook receivers, and the unauthenticated ops endpoints
    (`/api/v1/health`, `/api/v1/version`, `/api/v1/meta/enums`, `/api/v1/config`, `/metrics`,
    `/openapi.yaml`, `/docs`). Obtain an access token from `POST /api/v1/auth/login`;
    it is returned as `{ "accessToken", "expiresIn" }`. A long-lived refresh token is
    set as an HttpOnly cookie and rotated on every `POST /api/v1/auth/refresh`.

    **Authorization:** Two layers, both enforced server-side. A global role
    (`SUPER_ADMIN`, `ADMIN`, `USER`, `GUEST`) and a per-project membership role
    (`PROJECT_OWNER`, `PROJECT_ADMIN`, `PROJECT_MEMBER`, `PROJECT_VIEWER`). The
    effective project permission matrix is available at
    `GET /api/v1/projects/{projectId}/permissions`.

    **Error format:** All errors return
    `{ "code": "ERROR_CODE", "message": "...", "messageKey": "...", "details": {...} }`
    (`messageKey`/`details` optional).

    **Pagination:** List endpoints accept `?page=0&size=20` (zero-indexed pages,
    default size 20).

...
  contact:
    name: Octbase
...
servers:
  - url: /
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page: { name: page, in: query, required: false, schema: { type: integer, minimum: 0, default: 0 } }
    size: { name: size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 200, default: 20 } }
    projectId: { name: projectId, in: path, required: true, schema: { type: string, format: uuid } }
    taskId: { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
...
  responses:
...
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    Unprocessable: { description: Business-rule violation (e.g. LAST_OWNER, RELEASE_HAS_OPEN_TASKS), content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, example: VALIDATION_ERROR }
        message: { type: string }
        messageKey: { type: string }
        details: { type: object, additionalProperties: true }
...
    Preferences:
      type: object
      properties:
        userId: { type: string, format: uuid }
        language: { type: string, example: en }
        theme: { type: string, example: system }
...
paths:
...
  /api/v1/users/me/preferences:
    get:
      tags: [Dashboard]
      summary: Get my language/theme preferences
      operationId: getPreferences
      responses:
        '200': { description: Preferences, content: { application/json: { schema: { $ref: '#/components/schemas/Preferences' } } } }
    patch:
      tags: [Dashboard]
      summary: Update my language/theme preferences
      operationId: updatePreferences
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/Preferences' } } } }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Preferences' } } } }
        '422': { $ref: '#/components/responses/Unprocessable' }
...
  /api/v1/projects/{projectId}/activity:
    parameters: [ { $ref: '#/components/parameters/projectId' } ]
    get:
      tags: [Projects]
      summary: Project activity feed
      operationId: getProjectActivity
      parameters: [ { $ref: '#/components/parameters/page' }, { $ref: '#/components/parameters/size' } ]
      responses:
        '200': { description: Activity entries }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────── Memberships ─────────────────────
...
  /api/v1/tasks/{taskId}/activity:
    parameters: [ { $ref: '#/components/parameters/taskId' } ]
    get: { tags: [Tasks], summary: Task activity feed, operationId: getTaskActivity, responses: { '200': { description: Activity entries }, '404': { $ref: '#/components/responses/NotFound' } } }
...